refactor(api): Moving some files
