--- src/main/kotlin/com/higor/restaurantautomation/adapters/entity/Role.kt
@@ -1,6 +1,0 @@
-package com.higor.restaurantautomation.adapters.entity
-
-enum class Role {
-    USER,
-    ADMIN
-}
