+++ src/main/kotlin/com/higor/restaurantautomation/domain/model/user/Role.kt
@@ -1,0 +1,6 @@
+package com.higor.restaurantautomation.domain.model.user
+
+enum class Role {
+    USER,
+    ADMIN,
+}
