+++ src/main/kotlin/com/higor/restaurantautomation/domain/service/auth/AuthenticationService.kt
@@ -1,7 +1,7 @@
 package com.higor.restaurantautomation.domain.service.auth
 
-import com.higor.restaurantautomation.domain.dto.AuthenticationDtoIn
-import com.higor.restaurantautomation.domain.dto.AuthenticationDtoOut
+import com.higor.restaurantautomation.domain.dto.auth.AuthenticationDtoIn
+import com.higor.restaurantautomation.domain.dto.auth.AuthenticationDtoOut
 
 interface AuthenticationService {
     fun execute(request: AuthenticationDtoIn): AuthenticationDtoOut
