+++ src/main/kotlin/com/higor/restaurantautomation/domain/service/auth/RegisterService.kt
@@ -1,7 +1,7 @@
 package com.higor.restaurantautomation.domain.service.auth
 
-import com.higor.restaurantautomation.domain.dto.AuthenticationDtoOut
-import com.higor.restaurantautomation.domain.dto.RegisterDto
+import com.higor.restaurantautomation.domain.dto.auth.AuthenticationDtoOut
+import com.higor.restaurantautomation.domain.dto.auth.RegisterDto
 
 interface RegisterService {
     fun execute(request: RegisterDto): AuthenticationDtoOut
