+++ src/main/kotlin/com/higor/restaurantautomation/domain/service/auth/impl/AuthenticationServiceImpl.kt
@@ -2,8 +2,8 @@ package com.higor.restaurantautomation.domain.service.auth.impl
 
 import com.higor.restaurantautomation.configuration.security.JWTUtil
 import com.higor.restaurantautomation.configuration.security.UserDetailsImpl
-import com.higor.restaurantautomation.domain.dto.AuthenticationDtoIn
-import com.higor.restaurantautomation.domain.dto.AuthenticationDtoOut
+import com.higor.restaurantautomation.domain.dto.auth.AuthenticationDtoIn
+import com.higor.restaurantautomation.domain.dto.auth.AuthenticationDtoOut
 import com.higor.restaurantautomation.domain.service.auth.AuthenticationService
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
