+++ src/main/kotlin/com/higor/restaurantautomation/domain/service/company/GetCompanyByIdService.kt
@@ -1,6 +1,6 @@
 package com.higor.restaurantautomation.domain.service.company
 
-import com.higor.restaurantautomation.domain.model.CompanyModel
+import com.higor.restaurantautomation.domain.model.company.CompanyModel
 import java.util.UUID
 
 interface GetCompanyByIdService {
