+++ src/main/kotlin/com/higor/restaurantautomation/domain/service/user/GetAllUsersService.kt
@@ -1,8 +1,8 @@
 package com.higor.restaurantautomation.domain.service.user
 
 import com.higor.restaurantautomation.domain.dto.PaginationDto
-import com.higor.restaurantautomation.domain.dto.UserDtoOut
-import com.higor.restaurantautomation.domain.dto.UsersFilter
+import com.higor.restaurantautomation.domain.dto.user.UserDtoOut
+import com.higor.restaurantautomation.domain.dto.user.UsersFilter
 import org.springframework.data.domain.Pageable
 
 interface GetAllUsersService {
