+++ src/main/kotlin/com/higor/restaurantautomation/utils/extensions/UserModelExtension.kt
@@ -1,11 +1,11 @@
 package com.higor.restaurantautomation.utils.extensions
 
-import com.higor.restaurantautomation.adapters.entity.Role
 import com.higor.restaurantautomation.adapters.entity.User
-import com.higor.restaurantautomation.domain.dto.RegisterDto
-import com.higor.restaurantautomation.domain.dto.UserDtoIn
-import com.higor.restaurantautomation.domain.model.CompanyModel
-import com.higor.restaurantautomation.domain.model.UserModel
+import com.higor.restaurantautomation.domain.dto.auth.RegisterDto
+import com.higor.restaurantautomation.domain.dto.user.UserDtoIn
+import com.higor.restaurantautomation.domain.model.company.CompanyModel
+import com.higor.restaurantautomation.domain.model.user.Role
+import com.higor.restaurantautomation.domain.model.user.UserModel
 import com.higor.restaurantautomation.utils.PasswordHelper
 import java.time.Instant
 import java.util.UUID
