+++ src/main/resources/application.yml
@@ -7,7 +7,7 @@ spring:
             non_contextual_creation: 'true'
         format_sql: 'true'
     hibernate:
-      ddl-auto: update
+      ddl-auto: none
     show-sql: 'true'
     database-platform: org.hibernate.dialect.PostgreSQLDialect
   datasource:
