+++ src/test/kotlin/com/higor/restaurantautomation/controller/auth/AuthenticationControllerTest.kt
@@ -5,7 +5,7 @@ import com.higor.restaurantautomation.adapters.entity.Company
 import com.higor.restaurantautomation.adapters.entity.User
 import com.higor.restaurantautomation.adapters.repository.company.CompanyRepository
 import com.higor.restaurantautomation.adapters.repository.user.UserRepository
-import com.higor.restaurantautomation.domain.dto.AuthenticationDtoIn
+import com.higor.restaurantautomation.domain.dto.auth.AuthenticationDtoIn
 import com.higor.restaurantautomation.utils.PasswordHelper
 import com.higor.restaurantautomation.utils.factories.Factory
 import org.junit.jupiter.api.AfterEach
