+++ src/test/kotlin/com/higor/restaurantautomation/domain/service/company/impl/CreateCompanyServiceTest.kt
@@ -1,7 +1,7 @@
 package com.higor.restaurantautomation.domain.service.company.impl
 
 import com.higor.restaurantautomation.adapters.repository.company.CompanyRepository
-import com.higor.restaurantautomation.domain.model.CompanyModel
+import com.higor.restaurantautomation.domain.model.company.CompanyModel
 import com.higor.restaurantautomation.domain.service.company.CompanyExistsService
 import com.higor.restaurantautomation.domain.service.company.CreateCompanyService
 import com.higor.restaurantautomation.domain.service.exception.ApiException
