+++ src/test/kotlin/com/higor/restaurantautomation/domain/service/user/impl/CreateUserServiceTest.kt
@@ -1,8 +1,8 @@
 package com.higor.restaurantautomation.domain.service.user.impl
 
 import com.higor.restaurantautomation.adapters.repository.user.UserRepository
-import com.higor.restaurantautomation.domain.model.CompanyModel
-import com.higor.restaurantautomation.domain.model.UserModel
+import com.higor.restaurantautomation.domain.model.company.CompanyModel
+import com.higor.restaurantautomation.domain.model.user.UserModel
 import com.higor.restaurantautomation.domain.service.company.GetCompanyByIdService
 import com.higor.restaurantautomation.domain.service.exception.ApiException
 import com.higor.restaurantautomation.domain.service.user.CreateUserService
