+++ src/test/kotlin/com/higor/restaurantautomation/domain/service/user/impl/GetAllUsersServiceTest.kt
@@ -2,7 +2,7 @@ package com.higor.restaurantautomation.domain.service.user.impl
 
 import com.higor.restaurantautomation.adapters.entity.User
 import com.higor.restaurantautomation.adapters.repository.user.UserRepository
-import com.higor.restaurantautomation.domain.dto.UsersFilter
+import com.higor.restaurantautomation.domain.dto.user.UsersFilter
 import com.higor.restaurantautomation.utils.factories.Factory
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
